+++ hw2/src/stack_float.h
@@ -10,5 +10,5 @@
 # 1 "stack_float_gen.h"
 # 1 "gen/stack_gen.h" 1
 # 2 "stack_float_gen.h" 2
-typedef struct stack_float{ size_t size; list_float * List;}stack_float;stack_float * new_stack_float();void stack_float_push(stack_float *, float);void stack_float_pop(stack_float *);int stack_float_top(stack_float *);
+typedef struct stack_float{ size_t size; list_float * List;}stack_float;stack_float * new_stack_float();void stack_float_push(stack_float *, float);void stack_float_pop(stack_float *);float stack_float_top(stack_float *);
 #endif
