+++ hw2/src/stack_short.h
@@ -10,5 +10,5 @@
 # 1 "stack_short_gen.h"
 # 1 "gen/stack_gen.h" 1
 # 2 "stack_short_gen.h" 2
-typedef struct stack_short{ size_t size; list_short * List;}stack_short;stack_short * new_stack_short();void stack_short_push(stack_short *, short);void stack_short_pop(stack_short *);int stack_short_top(stack_short *);
+typedef struct stack_short{ size_t size; list_short * List;}stack_short;stack_short * new_stack_short();void stack_short_push(stack_short *, short);void stack_short_pop(stack_short *);short stack_short_top(stack_short *);
 #endif
